Finally fixed the server not receiving messages!
Solution: Moving the server instantiation back to the bottom of BattleshipServerGUI.

Why does this work? I'm not entirely sure. My best guess is that the connectionStatusLabel being passed in as null was causing the BattleshipServerGUI to not work properly.

In any case, I know this solution works just fine, so we can now progress!