Preliminary Main Menu and Instruction Panels
The main menu panels and instruction panels work.

The main menu connect to server does work, but the login is just a dummy until we can get the database working.

I've also added a LoginConfirmationData class for the server and client to pass around for determining when to do stuff.