Dataclass Revamp
Nothing functionally has changed since the last commit, but a stronger baseline for GUI updating has been implemented.

Using the custom JLabels classes in the BattleshipClient, PlayerData, and BattleshipServer classes.

The necessary data is still accessible, and the GUI's should be able to be updated because the custom JLabel classes are being used in more places, basically.